Crate xray

Provides a client interface for AWS X-Ray

Structs

Aws
Context information about the AWS environment this segment was run in

AwsOperation
Information about an AWS operation

Client
X-Ray daemon client interface

Ec2

Ecs

ElasticBeanstalk
Information about an Elastic Beanstalk environment. You can find this information in a file named /var/elasticbeanstalk/xray/environment.conf on the latest Elastic Beanstalk platforms.

Exception
Detailed representation of an exception

Header
Parsed representation of X-Amzn-Trace-Id request header

Http
Describes an http request/response cycle

Request
Information about a request.

Response
Information about a response.

Seconds
Represents fractional seconds since the epoch These can be derived from std::time::Duration and be converted too std::time::Duration

Segment
Description of an internal application operation which may be an extension of an external operation

Service
An object with information about your application.

Sql
Information about a SQL operation

StackFrame
A summary of a single operation within a stack trace

Subsegment
Record information about the AWS services and resources that your application accesses. X-Ray uses this information to create inferred segments that represent the downstream services in your service map.

Tracing

Enums

Annotation
A value type which may be used for filter querying

Cause
Represents the cause of an errror

Error

SegmentId
Unique identifier of an operation within a trace

TraceId
Coorelates a string of spans together

Type Definitions

Result
Type alias for Results which may return xray::Errors